requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
schedule>=1.2.0
python-dateutil>=2.8.0
fake-useragent>=1.4.0
tenacity>=8.2.0
python-Levenshtein>=0.21.0
